close #177
* For `sjp.lm()`, `sjp.lmer()`, `sjp.glm()` and `sjp.glmer()`, `type =
"pred"` and `type = "pred.fe"` now plot error bars for `show.ci = TRUE`
and a discrete variable on the x-axis.
* For `sjp.lm()`, `sjp.lmer()`, `sjp.glm()` and `sjp.glmer()`, the
`...`-ellipses argument now is also passed down to all errorbars-geoms
in prediction- and effect-plots, so you can now use the
`width`-argument to show the small stripes at the lower/upper end of
the error bars.
* Effect-plots (`type = "eff"`) now plot data points for discrete
variables on the x-axis.