change behavior of show.std (fixes #649)
If standardized coefficients are displayed (show.std = TRUE), the model is refitted. For mixed models this can result in different p-values and t-values as compared to the unstandardized coefficients. At the moment this can result in a mismatch between the t-value (coming from the standardized model) and the p-value (coming from the unstandardized model). The fix shows p-values and t-values for both unstandardized and standardized coefficients (if they are different).

Formatting of p-values is moved to a separate function to call it for both models.

Parameters string.std.p and string.std.stat are added.